back/requirements.txt
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
drf-yasg==1.17.0
apscheduler==3.6.3
sqlalchemy==1.3.13
psutil==5.7.0
pyalsaaudio==0.8.4
django-cors-headers==3.2.1